WIP Moved GraspCfg into config_types
